fix edge case with multiple edits in auto-fix
When multiple diagnostics propose a fix for the same piece of code, this
will break the naive auto-fix algorithm. Example:

```nix
lib:
{
  foo = with (lib); "hello";
}
```

Here, both redundant parentheses and redundant with will be removed.

Previously, this would lead to:

```
lib:
{
  foo =  hello";
}
```

Now, with this temporary fix, only the redundant parentheses will be
removed in the first step. And this stage, there will be an error
reported for the redundant with.

Running nixf-diagnose again, will then remove the redundant with.